+++ src/firecrest/main.py
@@ -18,9 +18,6 @@ from fastapi.exceptions import RequestValidationError
 
 from starlette.exceptions import HTTPException as StarletteHTTPException
 from firecrest.status.health_check.health_checker_cluster import ClusterHealthChecker
-from firecrest.status.health_check.health_checker_data_transfer import (
-    DataTransferHealthChecker,
-)
 from starlette_context import plugins
 from starlette_context.middleware import RawContextMiddleware
 
@@ -119,17 +116,6 @@ async def schedule_tasks(scheduler: AsyncScheduler):
             ClusterHealthChecker(cluster).check,
             IntervalTrigger(seconds=cluster.probing.interval_check),
             id=f"check-cluster-{cluster.name}",
-        )
-    if (
-        settings.data_operation.data_transfer
-        and settings.data_operation.data_transfer.probing
-    ):
-        await scheduler.add_schedule(
-            DataTransferHealthChecker(settings.data_operation.data_transfer).check,
-            IntervalTrigger(
-                seconds=settings.data_operation.data_transfer.probing.interval
-            ),
-            id="check-storage",
         )
     await scheduler.add_schedule(
         SSHClientDependency.prune_client_pools,